InvoiceBook_website/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
bcrypt==3.1.7
cffi==1.14.1
click==7.1.2
colorama==0.4.3
coverage==5.2.1
DateTime==4.3
distlib==0.3.1
Django==3.1
django-cors-headers==3.4.0
django-csp==3.6
django-feature-policy==3.4.0
django-filter==2.3.0
django-referrer-policy==1.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
entrypoints==0.3
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.7.0
iniconfig==1.0.1
isort==5.3.2
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.5.1
legacy==0.1.6
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
mysqlclient==2.0.1
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.1
pytest-cov==2.10.0
pytest-django==3.9.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
virtualenv==20.0.30
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.1.0
